[Fix] Implement caching support in make_sovits_dataset
- Also implement definite indexing shortcut in manifest.search
- CAVEAT: Not handling cases where subsongs are partially missing
- Cache hit inherently disabled for 22 'sud_vo_adv_live_IDOL_001_(start|end)-0#-01.acb's, which are rare adventure voice packs with one single sample
- libmp3lame seems to report 'Queue input is backward in time' for large file sets

+++ make_sovits_dataset.py
@@ -17,20 +17,32 @@ if __name__ == "__main__":
     parser = ArgumentParser(
         description="Create a dataset for training a voice cloning model"
     )
+
     parser.add_argument("character", type=str, help="Character name")
+
+    parser.add_argument("-o", "--output", type=str, default="", help="Output filename")
+    parser.add_argument("-f", "--format", type=str, default="wav", help="Output format")
+    parser.add_argument("-b", "--bitrate", type=int, default=128, help="Output bitrate")
+
     parser.add_argument(
         "-g", "--greedy", action="store_true", help="Search through all adventures"
     )
-    parser.add_argument("-o", "--output", type=str, default="", help="Output path")
-    parser.add_argument("-f", "--format", type=str, default="wav", help="Output format")
-    parser.add_argument("-b", "--bitrate", type=int, default=128, help="Output bitrate")
-    parser.add_argument("--cache-dir", type=str, default="tmp/", help="Temp directory")
+    parser.add_argument(
+        "-c", "--cache-dir", type=str, default=".sovits-cache/", help="Cache directory"
+    )
+    parser.add_argument(
+        "-k", "--keep-cache", action="store_true", help="Skip cache cleanup"
+    )
+
     args = parser.parse_args()
 
     # ------------------------------ SANITY CHECKS
 
+    logger.info("Fetching manifest...")
+    m = gom.fetch()
+
     if args.output == "":
-        args.output = f"sovits_dataset_{args.character}.{args.format}"
+        args.output = f"sovits_dataset_v{m.revision._get_canon_repr()}_{args.character}{"_greedy" if args.greedy else ""}.{args.format}"
 
     if not args.output.endswith(args.format):
         ext = Path(args.output).suffix[1:]
@@ -43,43 +55,61 @@ if __name__ == "__main__":
     if os.path.dirname(args.output):
         os.makedirs(os.path.dirname(args.output), exist_ok=True)
 
+    cache_active = False
     if os.path.exists(args.cache_dir):
         assert os.path.isdir(args.cache_dir), f"{args.cache_dir} is not a directory"
-        assert os.listdir(args.cache_dir) == [], f"{args.cache_dir} is not empty"
+        if os.listdir(args.cache_dir):
+            cache_active = True
     else:
         os.makedirs(args.cache_dir)
 
     # ------------------------------ CORE
 
-    logger.info("Fetching manifest...")
-    m = gom.fetch()
+    target = m.search(f"sud_vo.*{args.character}.*")
+    if args.greedy:
+        target += m.search(f"sud_vo_adv.*")
+    target = set([f.name for f in target])
+    if not target:
+        logger.warning(f"Found no voice samples for '{args.character}, aborting")
+        exit(1)
+    logger.success(f"Found {len(target)} voice samples for '{args.character}'")
 
-    logger.info(f"Downloading voice samples of '{args.character}'...")
+    if cache_active:
+        logger.info("Computing cache diff...")
+        target -= set(
+            [
+                str(
+                    Path(
+                        "_".join(f.split("_")[:-1])
+                        if f.startswith("sud_vo_adv_")
+                        else f
+                    ).with_suffix(".acb")
+                )
+                for f in os.listdir(args.cache_dir)
+            ]
+        )
+
+    logger.info(f"Downloading samples...")
     m.download(
-        f"sud_vo_adv.*" if args.greedy else f"sud_vo_adv.*{args.character}.*",
+        *sorted(list(target)),  # sort for logging
         path=args.cache_dir,
         categorize=False,
         convert_audio=True,
         audio_format="wav",  # avoid double compression
         unpack_subsongs=True,
     )
-    m.download(
-        f"sud_vo(?!_adv).*{args.character}.*",
-        path=args.cache_dir,
-        categorize=False,
-        convert_audio=True,
-        audio_format="wav",
-    )
-
-    assert os.listdir(args.cache_dir), f"Found no voice samples for '{args.character}'"
 
     logger.info("Making dataset...")
     with open(os.path.join(args.cache_dir, "filelist.txt"), "w") as fout:
         for f in os.listdir(args.cache_dir):
 
-            if f == "filelist.txt" or (
-                f.startswith("sud_vo_adv_")
-                and f.split("_")[-1].split("-")[0] != args.character
+            if (
+                f == "filelist.txt"
+                or args.character not in f
+                or (
+                    f.startswith("sud_vo_adv_")
+                    and f.split("_")[-1].split("-")[0] != args.character
+                )
             ):
                 continue
                 # exclude other characters in adventure voice pack
@@ -93,6 +123,9 @@ if __name__ == "__main__":
         check=True,
     )
 
-    logger.success(f"Dataset saved to {args.output}")
-    logger.info("Cleaning up temporary files...")
-    shutil.rmtree(args.cache_dir)
+    os.remove(os.path.join(args.cache_dir, "filelist.txt"))
+    if not (cache_active or args.keep_cache):
+        logger.info("Purging cache...")
+        shutil.rmtree(args.cache_dir)
+
+    logger.success(f"Dataset ready at '{args.output}'")
